+++ apps/user/src/app/login/page.tsx
@@ -3,7 +3,7 @@
 import styled from 'styled-components';
 import Image from 'next/image';
 import BaseLayout from '@/layouts/BaseLayout';
-import useLogin from '@/services/auth/hooks/useLogin';
+import useLogin from './hook';
 import { useRouter } from 'next/navigation';
 import { SIGNUP_PAGE_ROUTE } from '@/constants/routes';
 import { Button, Column, Input, PreviewInput } from '@maru/ui';
