+++ apps/user/src/app/page.tsx
@@ -4,7 +4,7 @@ import MainLayout from '@/layouts/MainLayout';
 import Notice from '@/components/main/Notice';
 import Schedule from '@/components/main/Schedule';
 import initMockAPI from '@/mocks';
-import SchoolRecruitCard from '@/components/main/SchoolRecruitCard';
+import Dday from '@/components/main/Dday';
 import Faq from '@/components/main/Faq';
 import { Row } from '@maru/ui';
 import styled from 'styled-components';
@@ -19,7 +19,7 @@ const Home = () => {
         <MainLayout>
             <StyledMainPage>
                 <Row gap="48px" height="451px" width="100%" justifyContent="center">
-                    <SchoolRecruitCard />
+                    <Dday />
                     <Schedule />
                 </Row>
                 <Row gap="48px" width="100%" height="242px" justifyContent="center">
