+++ apps/user/src/components/signup/Terms.tsx
@@ -3,50 +3,15 @@ import { Dispatch, SetStateAction, useEffect, useState } from 'react';
 import { color, font } from '@maru/theme';
 import { flex } from '@maru/utils';
 import styled from 'styled-components';
+import useTerms from './hooks/Terms.hook';
 
 interface PropsType {
     setCheckTermsAgree: Dispatch<SetStateAction<boolean>>;
 }
 
 const Terms = ({ setCheckTermsAgree }: PropsType) => {
-    const [allCheck, setAllCheck] = useState(false);
-    const [checkOne, setChecktOne] = useState(false);
-    const [checkTwo, setChecktTwo] = useState(false);
-
-    const allCheckEvent = () => {
-        if (allCheck === false) {
-            setChecktOne(true);
-            setChecktTwo(true);
-        } else {
-            setChecktOne(false);
-            setChecktTwo(false);
-        }
-    };
-
-    const checkOneEvent = () => {
-        if (checkOne === false) {
-            setChecktOne(true);
-        } else {
-            setChecktOne(false);
-        }
-    };
-
-    const checkTwoEvent = () => {
-        if (checkTwo === false) {
-            setChecktTwo(true);
-        } else {
-            setChecktTwo(false);
-        }
-    };
-
-    useEffect(() => {
-        if (checkOne === true && checkTwo === true) {
-            setAllCheck(true);
-            setCheckTermsAgree(true);
-        } else {
-            setAllCheck(false);
-        }
-    }, [checkOne, checkTwo]);
+    const { allCheck, checkOne, checkTwo, allCheckEvent, checkOneEvent, checkTwoEvent } =
+        useTerms(setCheckTermsAgree);
 
     return (
         <StyledTerms>
